Break up index.js into functions
This allows us to add unit tests for the distinct parts.
